requirements.txt
torch
torchvision
numpy
opencv-python
pandas
pyyaml
pillow
tensorboard
scikit-learn
matplotlib
